fix some bugs

+++ application/portal/controller/CategoryBase.php
@@ -53,6 +53,10 @@ class CategoryBase extends Controller
         }
         (new IDPositive())->goCheck();
         $id=$request->param('id');
+        $count=CategoryModel::where('parent_id','=',$id)->count();
+        if($count>0){
+            return ResultService::failure('栏目下存在子栏目,无法删除');
+        }
         $category=CategoryModel::where('id','=',$id)->find();
         if($category){
             $category->delete();
@@ -89,7 +93,7 @@ class CategoryBase extends Controller
                 $category->thumbnail=$request->param('thumbnail');
             }
             if($request->has('more')){
-                $category->more=$request->param('more','','json_decode');
+                $category->more=json_decode($request->param('more'),true);
             }
             if($request->has('parent_id')){
                 if($request->param('parent_id')==$category->id){
@@ -121,6 +125,14 @@ class CategoryBase extends Controller
             (new IDPositive())->goCheck();
             $id=$request->param('id');
             $category=CategoryModel::where('id','=',$id)->with('template')->find();
+            if($category->parent_id==0){
+                $category->parent_name='一级栏目';
+            }
+            else{
+                $parent=CategoryModel::where('id','=',$category->parent_id)->find();
+                $category->parent_name=$parent?$parent->name:'';
+
+            }
             if($category){
                 $category->hidden(['create_time','update_time','template.create_time','template_update_time']);
                 return ResultService::makeResult(ResultService::Success,'',$category->toArray());
@@ -135,6 +147,26 @@ class CategoryBase extends Controller
         }
     }
 
+    /**分页获取栏目
+     * @param Request $request
+     * @return \think\response\Json
+     * @throws TokenException
+     */
+    public function getByPage(Request $request){
+        if(!TokenService::validAdminToken($request->header('token'))){
+            throw new TokenException();
+        }
+        $pageResult=[];
+        $category=CategoryModel::with('template')->select()->toArray();
+        $pageResult['total']=count($category);
+        if($request->has('page')){
+            $pageResult['page']=$page=$request->param('page');
+            $pageResult['pageSize']=$pageSize=$request->has('pageSize')?$request->param('pageSize'):config('base.defaultPageSize');
+            $category=CategoryModel::with('template')->page($page,$pageSize)->select();
+        }
+        $pageResult['pageData']=$category;
+        return ResultService::makeResult(ResultService::Success,'',$pageResult);
+    }
     /**获取栏目树
      * @param Request $request
      * @return \think\response\Json
@@ -144,7 +176,8 @@ class CategoryBase extends Controller
         if(!TokenService::validAdminToken($request->header('token'))){
             throw new TokenException();
         }
-        $tag=$request->has('tag')?$request->param('tag'):'checked';//表示权限是否被允许的键名，用于前端生成带复选框的树时使用
+        $type=$request->has('type')?$request->param('type'):0;//返回结果的类型 0为树型 1为表格树型
+        $tag=$request->has('tag')?$request->param('tag'):'checked';//表示是否被允许的键名，用于前端生成带复选框的树时使用
         $tagSuccessValue=$request->has('tagSuccessValue')?$request->param('tagSuccessValue'):'1';
         $tagFailureValue=$request->has('tagFailureValue')?$request->param('tagSuccessValue'):'0';
         if($request->has('postId')){
@@ -155,14 +188,19 @@ class CategoryBase extends Controller
             foreach($categories as $item){
                 $postCategory[]=$item['category_id'];
             }
-            return ResultService::makeResult(ResultService::Success,'',CategoryModel::generateCategoryTree($postCategory,$tag,$tagSuccessValue,$tagFailureValue));
+            $tree=CategoryModel::generateCategoryTree($postCategory,$tag,$tagSuccessValue,$tagFailureValue);
         }
         else{
+            $tree=CategoryModel::generateCategoryTree(array(),$tag,$tagSuccessValue,$tagFailureValue);
             //var_dump(CategoryModel::generateCategoryTree(array(),$tag,$tagSuccessValue,$tagFailureValue));
-            return ResultService::makeResult(ResultService::Success,'',CategoryModel::generateCategoryTree(array(),$tag,$tagSuccessValue,$tagFailureValue));
+        }
+        if($type==0){
+            return ResultService::makeResult(ResultService::Success,'',$tree);
+        }
+        else{
+            return ResultService::makeResult(ResultService::Success,'',CategoryModel::generateCategoryTreeTable($tree,0));
         }
     }
-
     /**获取栏目下内容
      * @param Request $request
      * @return \think\response\Json
